+++ backend/tests/app/test_dependencies.py
@@ -81,11 +81,12 @@ def test_get_connection_string_full(monkeypatch, patch_required_env):
     ), "must return fully formed connection string"
 
 
-def test_get_starting_agent(patch_required_env):
+def test_get_starting_agent(patch_required_env, get_weather_tool):
     settings = Settings()
-    agent = get_starting_agent(settings)
+    agent = get_starting_agent(settings, tool_list=[get_weather_tool])
 
     assert isinstance(agent, Agent)
+    assert agent.tools == [get_weather_tool]
 
 
 @pytest.mark.asyncio